GUI/requirements.txt
flask
pyserial
Flask-Cors
requests
playsound
pydub